+++ src/apexpy/helpers.py
@@ -146,7 +146,7 @@ def subsol(datetime):
 
     Parameters
     ==========
-    datetime : :class:`datetime.datetime`
+    datetime : :class:`datetime.datetime` or :class:`numpy.ndarray[datetime64]`
 
     Returns
     =======
@@ -171,21 +171,32 @@ def subsol(datetime):
 
     """
     # convert to year, day of year and seconds since midnight
-    year = datetime.year
-    doy = datetime.timetuple().tm_yday
-    ut = datetime.hour * 3600 + datetime.minute * 60 + datetime.second
+    if isinstance(datetime, dt.datetime):
+        year = np.asanyarray([datetime.year])
+        doy = np.asanyarray([datetime.timetuple().tm_yday])
+        ut = np.asanyarray([datetime.hour * 3600 + datetime.minute * 60 + datetime.second])
+    elif isinstance(datetime, np.ndarray):
+        times = datetime.astype('datetime64[s]')  # works for datetime of wrong precision or unix epoch
+        year_floor = times.astype('datetime64[Y]')
+        day_floor = times.astype('datetime64[D]')
+        year = year_floor.astype(int) + 1970
+        doy = (day_floor - year_floor).astype(int) + 1
+        ut = (times.astype('datetime64[s]') - day_floor).astype(float)
+    else:
+        raise ValueError("input must be datetime.datetime or numpy array")
 
-    if not 1601 <= year <= 2100:
+    if not (np.all(1601 <= year) and np.all(year <= 2100)):
         raise ValueError('Year must be in [1601, 2100]')
 
     yr = year - 2000
 
-    nleap = int(np.floor((year - 1601.0) / 4.0))
+    nleap = np.floor((year - 1601.0) / 4.0).astype(int)
     nleap -= 99
-    if year <= 1900:
-        ncent = int(np.floor((year - 1601.0) / 100.0))
+    mask_1900 = year <= 1900
+    if np.any(mask_1900):
+        ncent = np.floor((year[mask_1900] - 1601.0) / 100.0).astype(int)
         ncent = 3 - ncent
-        nleap = nleap + ncent
+        nleap[mask_1900] = nleap[mask_1900] + ncent
 
     l0 = -79.549 + (-0.238699 * (yr - 4.0 * nleap) + 3.08514e-2 * nleap)
     g0 = -2.472 + (-0.2558905 * (yr - 4.0 * nleap) - 3.79617e-2 * nleap)
@@ -215,12 +226,15 @@ def subsol(datetime):
 
     # Equation of time (degrees):
     etdeg = lmean - alpha
-    nrot = round(etdeg / 360.0)
+    nrot = np.round(etdeg / 360.0)
     etdeg = etdeg - 360.0 * nrot
 
-    # Subsolar longitude:
-    sslon = 180.0 - (ut / 240.0 + etdeg) # Earth rotates one degree every 240 s.
-    nrot = round(sslon / 360.0)
+    # Subsolar longitude calculation. Earth rotates one degree every 240 s.
+    sslon = 180.0 - (ut / 240.0 + etdeg)
+    nrot = np.round(sslon / 360.0)
     sslon = sslon - 360.0 * nrot
 
+    # Return a single value from the output if the input was a single value
+    if isinstance(datetime, dt.datetime):
+        return sslat[0], sslon[0]
     return sslat, sslon